Fixed new warnings from r282.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@284 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/HLE/HLE.cpp
@@ -103,7 +103,7 @@ void PatchFunctions()
 		{
             const Debugger::CSymbol& rSymbol = Debugger::GetSymbol(SymbolIndex);
 			u32 HLEPatchValue = (1 & 0x3f) << 26;
-			for (int addr = rSymbol.vaddress; addr < rSymbol.vaddress + rSymbol.size; addr+=4) {
+			for (size_t addr = rSymbol.vaddress; addr < rSymbol.vaddress + rSymbol.size; addr+=4) {
 				Memory::Write_U32(HLEPatchValue | i, addr);
 			}
 			//PanicAlert("patched %s", OSPatches[i].m_szPatchName);
